refactor: move verify refresh token util to shared utils folder
